nearby-poi: compare high prec locations in removeSimpleDuplicates() to handle duplicate POI which were moved to different places by WrongAngleFixer.

+++ trunk/src/uk/me/parabola/imgfmt/Utils.java
@@ -387,7 +387,7 @@ public class Utils {
 		int latHp = co.getHighPrecLat();
 		int lonHp = co.getHighPrecLon();
 		
-		return (long)(latHp & 0xffffffffL) << 32 | (lonHp & 0xffffffffL);
+		return (latHp & 0xffffffffL) << 32 | (lonHp & 0xffffffffL);
 	}
 	
 	public static int numberToPointerSize(int n) {
